refactor(subscriptions): extract section subscriptions feature
