+++ src/lib/mcp/tools/calendar-subscription-tool-types.ts
@@ -1,5 +1,5 @@
 import type { AuthInfo } from "@modelcontextprotocol/sdk/server/auth/types.js";
-import type { getUserCalendarSubscription } from "@/features/home/server/subscription-read-model";
+import type { getUserCalendarSubscription } from "@/features/subscriptions/server/subscription-read-model";
 import type { resolveMcpMode } from "@/lib/mcp/tools/_helpers";
 
 export type ToolExtra = { authInfo?: AuthInfo };
